refactor(nvim-*): remove 'nvim' from filename
